This is the big one. So glad we have version control ;)
This patch does a number of thigs. A specialized class TaskHierarchy
aids in building (drum roll...) the task hierarcht. The renderer is
simplified, and all data to draw the menu is pulled from the database
in a single query; hitting the database in a nested loop is a draw on
performance, and this approach allows for easier task menu extensions.
I've added 'my stories' and 'In progress' for starters.

The task hierarchy now lives in a fold-out menu; in the previous
implementation the task list was always either over- or underlapping
the task entry screen. That should be gone now.



git-svn-id: http://agilito.googlecode.com/svn/trunk@36 573662cc-8f25-11dd-82cb-3113a4e48fd9